+++ samples/15.handling-attachments/pom.xml
@@ -41,7 +41,7 @@
         <java.version>1.8</java.version>
         <maven.compiler.target>1.8</maven.compiler.target>
         <maven.compiler.source>1.8</maven.compiler.source>
-        <start-class>com.microsoft.bot.sample.echo.Application</start-class>
+        <start-class>com.microsoft.bot.sample.attachments.Application</start-class>
     </properties>
 
     <dependencies>
@@ -125,7 +125,7 @@
                               <goal>repackage</goal>
                           </goals>
                           <configuration>
-                              <mainClass>com.microsoft.bot.sample.echo.Application</mainClass>
+                              <mainClass>com.microsoft.bot.sample.attachments.Application</mainClass>
                           </configuration>
                       </execution>
                   </executions>
